@@ -1,30 +1,30 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.13'
+gem 'rails', '4.0.0'
 
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails',   '~> 4.0.0'
+  gem 'coffee-rails', '~> 4.0.0'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   gem 'therubyracer', :platforms => :ruby
 
-  gem 'uglifier', '>= 1.0.3'
+  gem 'uglifier', '>= 1.3.0'
 end
 
 gem 'jquery-rails'
 
 
-gem 'sufia', github: 'projecthydra/sufia', ref: '2c8a438' #'~> 2.1.0'
+gem 'sufia', '~> 3.0.0'
 #gem 'sufia', path: '../sufia'
 #gem 'hydra-pbcore', github:'curationexperts/hydra-pbcore', branch: 'devel'
 #gem 'hydra-pbcore', :path=>'../hydra-pbcore'
-gem 'rdf', github: 'ruby-rdf/rdf', ref: '9b20f6d' # Keep until rdf 1.0.8 is released
+gem 'rdf', github: 'ruby-rdf/rdf', ref: '9b20f6d' # Keep until rdf 1.0.8 (or 1.1.0) is released
 gem 'hydra-pbcore', '2.2.0rc1'
-gem 'rails_admin', "~> 0.4.5"
+gem 'rails_admin', "~> 0.5.0"
 
 group :development, :test do
   #gem 'byebug' # This is a debugger, can be removed when not needed by the developer
