@@ -11,6 +11,8 @@ class GenericFile < ActiveFedora::Base
 
   delegate_to 'descMetadata', [:has_location, :program_title, :series_title,
                                :item_title, :episode_title, :has_event,
+                               :event_location, :production_location,
+                               :filming_event, :production_event,
                                :has_event_attributes,
                                :creator_attributes, :contributor_attributes, 
                                :publisher_attributes, :has_location_attributes,
@@ -21,8 +23,8 @@ class GenericFile < ActiveFedora::Base
   attr_accessible  :part_of, :contributor_attributes, :creator_attributes,
         :title_attributes, :description_attributes, :publisher_attributes,
         :date_created, :date_uploaded, :date_modified, :subject, :language,
-        :rights, :resource_type, :identifier, :has_location_attributes, :tag,
-        :related_url, :permissions
+        :rights, :resource_type, :identifier, :event_location,
+        :production_location, :tag, :related_url, :permissions
 
   before_destroy :remove_content
 
@@ -35,8 +37,24 @@ class GenericFile < ActiveFedora::Base
     publisher.build if publisher.empty?
     contributor.build if contributor.empty?
     creator.build if creator.empty?
-    has_location.build if has_location.empty?
+    #has_location.build if has_location.empty?
     description.build if description.empty?
+    super
+  end
+
+  
+  def remove_blank_assertions
+    publisher.select { |p| p.name.first == '' && p.role.first == ''}.each(&:destroy)
+    contributor.select { |p| p.name.first == '' && p.role.first == ''}.each(&:destroy)
+    creator.select { |p| p.name.first == '' && p.role.first == ''}.each(&:destroy)
+    # events (filming events and production events specifically) must have locations
+    has_event.each do |event|
+      event.has_location.each do |location|
+        location.destroy if location.location_name.first == ''
+      end
+    end
+    description.select { |p| p.value.first == '' && p.type.first == ''}.each(&:destroy)
+    title.select { |p| p.value.first == '' && p.title_type.first == ''}.each(&:destroy)
     super
   end
 
@@ -114,9 +132,10 @@ class GenericFile < ActiveFedora::Base
   end
 
   def terms_for_display
-    [ :part_of, :contributor, :creator, :title, :description, 
+    [ :part_of, :contributor, :creator, :title, :description, :event_location, :production_location,
         :publisher, :date_created, :date_uploaded, :date_modified, :subject, :language, :rights, 
-        :resource_type, :identifier, :has_location, :tag, :related_url]
+        :resource_type, :identifier, :tag, :related_url]
+    # :has_location
   end
   
   ## Extract the metadata from the content datastream and record it in the characterization datastream
@@ -142,6 +161,8 @@ class GenericFile < ActiveFedora::Base
     descMetadata.has_location #.map(&:location_name).flatten
   end
 
+
+  # Necessary because parts of sufia call creator= with a string.
   ### Map creator[] -> creator[].name
   # @param [Array,String] creator_properties a list of hashes with role and name or just names
   def creator=(args)
@@ -149,9 +170,15 @@ class GenericFile < ActiveFedora::Base
       raise ArgumentError, "You must provide a string or an array.  You provided #{args.inspect}"
     end
     args = Array(args)
-    self.creator_attributes = [{name: args, role: "Uploader"}]
+    if args.first.is_a?(String)
+      return if args == [''] 
+      self.creator_attributes = [{name: args, role: "Uploader"}]
+    else
+      descMetadata.creator = args
+    end
   end
 
+  # Necessary because parts of sufia call title= with a string.
   ### Map title[] -> title[].value
   # @param [Array,String] title_properties a list of hashes with type and value
   def title=(args)
@@ -159,20 +186,11 @@ class GenericFile < ActiveFedora::Base
       raise ArgumentError, "You must provide a string or an array.  You provided #{args.inspect}"
     end
     args = Array(args)
-    self.title_attributes = [{name: args, title_type: "Program"}]
-  end
-
-  ### Map based_near[] -> has_location[].locationName
-  # @param [Array] vals a list of hashes with location_name
-  def has_location=(vals)
-    existing = descMetadata.has_location
-    descMetadata.has_location = [] if existing.size > vals.size
-    Array(vals).each_with_index do |val, index|
-      obj = descMetadata.has_location[index]
-      if obj.nil?
-        obj = descMetadata.has_location.build
-      end
-      obj.location_name = val['location_name']
+    if args.first.is_a?(String)
+      return if args == [''] 
+      self.title_attributes = [{name: args, title_type: "Program"}]
+    else
+      descMetadata.title=args
     end
   end
 
@@ -184,7 +202,10 @@ class GenericFile < ActiveFedora::Base
     self.creator.each { |c| c.destroy } if params[:creator_attributes]
     self.contributor.each { |c| c.destroy } if params[:contributor_attributes]
     self.producer.each { |c| c.destroy } if params[:producer_attributes]
+    self.publisher.each { |c| c.destroy } if params[:publisher_attributes]
     self.title.each { |c| c.destroy } if params[:title_attributes]
+    self.event.each { |c| c.destroy } if params[:event_attributes]
+    self.description.each { |c| c.destroy } if params[:description_attributes]
   end
 
 
@@ -194,9 +215,10 @@ class GenericFile < ActiveFedora::Base
   end
 
   def to_pbcore_xml
-    doc = HydraPbcore::Datastream::Document.new
-    doc.title = program_title
-    doc.alternative_title = series_title
+    doc = ExportPbcoreDatastream.new 
+    title.each do |t|
+      doc.title(t.title_type, t.value)
+    end
     descMetadata.creator.each do |c|
       doc.insert_creator c.name.first, c.role.first
     end
@@ -206,9 +228,9 @@ class GenericFile < ActiveFedora::Base
     descMetadata.publisher.each do |c|
       doc.insert_publisher c.name.first, c.role.first
     end
-    descMetadata.has_location.each do |l|
-      doc.insert_place l.location_name.first
-    end
+    # descMetadata.has_location.each do |l|
+    #   doc.insert_place l.location_name.first
+    # end
 
     doc.insert_date(date_created.first)
     doc.asset_type = resource_type.to_a
