@@ -37,11 +37,10 @@ class ImportedMetadata < ActiveFedora::Base
      'description_attributes' => [
        {'type' => 'Log', 'value' => description},
      ],
-     'has_location_attributes' => [
-       {'location_name' => event_location},
-     ],
      'has_event_attributes' => [
-       {'date_time' => date_portrayed},
+       {'has_location_attributes' => [{'location_name' => event_location}],
+        'event_definition' => 'Filming', 
+        'date_time' => date_portrayed},
      ],
      'applied_template_id' => pid, 'unarranged' => 'false'
     }.with_indifferent_access
