@@ -11,9 +11,6 @@ Bawstun::Application.configure do
   config.serve_static_assets = true
   config.static_cache_control = "public, max-age=3600"
 
-  # Log error messages when you accidentally call methods on nil
-  config.whiny_nils = true
-
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
@@ -29,9 +26,6 @@ Bawstun::Application.configure do
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
 
-  # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
-
   # Print deprecation notices to the stderr
   config.active_support.deprecation = :stderr
 
@@ -41,4 +35,6 @@ Bawstun::Application.configure do
   # Base directory for original content storage 
   config.external_store_base = File.expand_path('../../../spec/storage', __FILE__)
   config.storage_manager = 'NullStorageManager'
+
+  config.eager_load = false
 end
